Organized the folder directory
