+++ src/components/Index.js
@@ -1,0 +1,3 @@
+import CustomButton from "./CustomButton";
+
+export {CustomButton}
