+++ src/constants/CommonStyles.js
@@ -1,0 +1,24 @@
+import {StyleSheet} from 'react-native';
+
+const CommonStyles = StyleSheet.create({
+  dFlex: {
+    flex: 1,
+  },
+  flexRow: {
+    flexDirection: 'row',
+  },
+  flexColumn: {
+    flexDirection: 'column',
+  },
+  p10: {
+    padding: 10,
+  },
+  m10: {
+    margin: 10,
+  },
+  centerItems: {
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
+export default CommonStyles;
