+++ src/constants/Images.js
@@ -1,0 +1,5 @@
+const Images = {
+  homeImage: require('../assets/calculatorImage.jpg'),
+};
+
+export default Images;
